Sorted the Images folder
The new folder structure looks like this:

> Images
>> Example Assets
>>> backgrounds
>>> portraits
>>> next-indicator
>> Plugin
>>> Context Menus
>>> Event Icons
>>>> Main Icons
>>> Pieces
>>> Resources
>>> Toolbar
>>> Tutorials

There are some images in the Images folder but I think they aren't used at all.  I left them there in case there is a plan I don't know of. Otherwise they can be deleted

I also duplicated the X - image (for the set-value event), because the Event was using the one from the resources folder. There was also an Icon called set-value that was somehow not used. It is renamed to set-value old now.

On that note, I renamed all the event icons to have the exact name of their event.

I also corrected all places inside scripts that I could find. The rest should have corrected itself, just caused a lot of changes in a lot of the files.