Image organization and creation of the Example Assets folder (#183)
* Sorted the Images folder

The new folder structure looks like this:

> Images
>> Example Assets
>>> backgrounds
>>> portraits
>>> next-indicator
>> Plugin
>>> Context Menus
>>> Event Icons
>>>> Main Icons
>>> Pieces
>>> Resources
>>> Toolbar
>>> Tutorials

There are some images in the Images folder but I think they aren't used at all.  I left them there in case there is a plan I don't know of. Otherwise they can be deleted

I also duplicated the X - image (for the set-value event), because the Event was using the one from the resources folder. There was also an Icon called set-value that was somehow not used. It is renamed to set-value old now.

On that note, I renamed all the event icons to have the exact name of their event.

I also corrected all places inside scripts that I could find. The rest should have corrected itself, just caused a lot of changes in a lot of the files.

* Fixing Image Paths

Turns out I forgot to fix some.
Now it should work ;).

* Changed the folder structure once again

- Now there is an Example Assets folder containing the fonts, backgrounds, etc.
- Because of this I moved all of the folders back from Images/Plugin into Images.

* Deleted images that were not needed

Deleted images that were not needed. Searched through the scripts for References and checked their owners. None.

* Fix the fonts

Fixed the references to the fonts inside the fonts...

Yeah, hope it works now.

* Fix the Audio Icons problem

They are replaced by the editor icons now.

The icons are deleted now.

* replacing the remove clear icon with a native icon

* adding path replacer and using more theme icons

Co-authored-by: Emilio Coppola <coppolaemilio@gmail.com>