Mirror portraits setting for characters and the Join event (updated) (#208)
* Adding Mirror Settings for Positions right and center_right for a character

I did the mirroring by setting the rect_scale.x of the texture-rect to -1. To make it so the character isn't moved by this, also the TextureRects pivot get's set to the horizontal center of the TextureRect.

* Make the mirror option a lot easier

Turns out there is a flip value for TextureRect. Use it now.

* Add the requested features to the Mirror Option

Moved the characters mirror option and made it so it mirrors all portraits on all positions.

The Join Event now has a mirror toggle.

* Updated the Readme