New way of integrating the docs
There now exist a couple of nodes to use in your plugin:
- DocsPageViewer (A RichTextLabel that can parse Markdown and handle links, scroll to sections, etc.)
- DocumentationTree (A tree with all the items of the Docs Content folder)
- DocumentationViewer (A working combination of the DocsPageViewer and the DocumentationTree)
You can now just use the DocumentationViewer as a kind of Documentation tab or page in your plugin. Or you can integrate the DocsPageViewer into an exisiting editor. It is also very easy to add the tree items to an existing tree.

This commit also worked on to the integration into dialogic, but I try to keep it as minimal as possible.