Adding Nicknames (New Feature) and fixing auto-color names in text events (#246)
* changed character editor to include a nickname checkbox and line edit

* updated character list to include nicknames

* fixed an issue with name color and added nicknames

This only changes the parse_characters function. I fixed an issue in which names would be colored (when having auto-color) while they were in another word. e.g. May and Maybe. May should not be auto-colored while part of Maybe or another name like Mayleigh but it was. Having it check the character after for a space or punctuation alleviates this issue (this also assumes people capitalize the first letter of a name). In complex languages like like Japanese with 3+ alphabets or any other character based languages without capital letters, this still helps A LOT with that issue but is not an airtight solution. Additionally, since multiple Nicknames can be added in the character editor with just a comma to separate in the line edit, it needed to be broken up and then the same logic applies as the name as seen in the changes.