Add an event_id system to dialogic (#268)
* Add an event_id system to dialogic

There are a few big changes in this pr:
- Every event now has to have a 'event_id' key!!!
   - it looks something like this: 'dialogic_003'
   - for each of the sections I decided to start from the next multiple of ten. This would allow fitting in more event inbetween, altrough there would be no functionality difference if the numbers weren't ordered.
- This is used when building the timelines, and when parsing them in game.
- I renamed a couple of events (their scenes and scripts) to fit their actual names:
   - TextBlock -> TextEvent
   - IFCondition -> Condition
   - CharacterJoinBlock/CharacterLeaveBlock -> CharacterJoin/CharacterLeave
Unfortunately there was some crash with the branch parser, so I had to rename a couple of variables and keys there. If you don't like them, please tell me.

I also updated the defaule Error text to have an id.

And I also updated the theme
AND of course the big problem of HOW TO UPDATE THE TIMELINES.... I added a function that will check all timelines and their events and add the id's if needed. Unfortunately I didn't find a way to only call this once, so now it's called everytime the editor is started... Please tell me how to do it right, if you know!

If this pr gets merged there are a couple of other things that can be done like deleting pointless keys (e.g the close_dialog key in the close_dialog event was only used to identify the event)

* making it identical to main

* Event Fixer rework

It is now called resouce_fixer. Also there is now a update_number in the settings.cfg file. It will be used to find out what updates need to be done to the resources for them to work with anything we change about dialogic.

* Fixed TextEvent Reference

Co-authored-by: Emilio Coppola <coppolaemilio@gmail.com>