Add SetGlossary event (#290)
* Fixed Dependecies

I moved some scenes in the last minute. Stupid me. I hope it is fixed now.

* Revert "Fixed Dependecies"

This reverts commit 9e7bcd8dbb29b69aa6373ad5553ebd686f437de6.

* Add SetGlossary Event

Mostly done to test how hard it would be to add a new event. It was very easy. Had to also implement a new set_glossary_by_id() function to the DialogicSignleton. Also the definitions are reloaded (in the dialog node) everytime the text is updated now.  

A small change was the addition of a max_preview_characters variable in the ExpandControl (that also shows the preview).