Update dialog_node.gd
This way you can warn the user that he did not add his node to a canvas layer and that it probably won't show up in the game as expects. 

It is not an automatic way to solve the problem, but a way to tell you what the problem is.