Add the new styles to the remaining pickers
TimelinePicker, ThemePicker, OperationPicker, GlossaryPicker. It also removes the [] around the items, except for the operation picker. I wasn't sure if it wouldn't make readability harder having three similar buttons next to each other, so I kept the [] there. 

And it adds a "Bus:" label to the AudioPicker, because I thought it would make it cleaner.