Work on the Resource Folders
Right now folders and subfolders work for the timelines (because I wanted to get it working before I expand it). 

There is a new file called "folder_structure.json" in the dialogic folder. It is parsed to create the tree. 

Most of the functions to work with the folder structure are in the DialogicUtil.

I have moved the Popups into the MasterTree. They are now created in the script.
There is a new "RemoveFolderConfrimation" popup.

Also I changed the interaction between the MasterTree and the TimelineEditor for creating a new timeline. All the Tree stuff is now done in the tree ('new_timeline()') and only the create_timeline function of the TimelineEditor is called.

Next I want to try working on dragging and dropping resources. We'll see how that goes. After that there needs to be a lot of cleanup, because many things are not very nice right now. Wish me luck!