Implement and polish ResourceFolders
ResourceFolders are now working for all 4 resource types. 

Like for the timelines, the Popupmenus have been expanded and the creation of new resources was shifted into the MasterTree.gd. 

Resources are now really deleted when one of their parent folders is deleted.

Folders now have metadata. Right now it contains a color key and a folded key. The second one is used to save the folded/unfolded state of each folder. The color key is not used right now. Metadata can easily be read and written with two easy functions in the DialogicUtil.

When moving, adding and renaming folders, it now makes sure that no folder folder whith that name already exists (as that produced errors).

The folder structure is now updated once when the editor is loaded to make sure there are no files missing.

I'm not sure what the problem was, but when editing the names in the CharacterEditor or the DefinitionEditor, the LineEdit lost focus each time you entered a character. This was fixed by just grabbing the focus back ;).

Btw, something I didn't explain in the previous commits: The master tree has some new functions that are used to get the path to an item in the tree. That path can be used to access for example the folder in the folder structure.