Small Bug fixes
When dragging one of the root folders, the folder structure would become corrupted. Not anymore.

When editing a folder name but not changing it, it would corrupt the folder. Not anymore.

Turns out moving folders didn't work anymore, since it removes the old folder and then adds the new one. But when removing the old one it also REALLY deleted all the contained files. So each time you moved a folder it would delete all it's files. 

Took me long to find out and short to fix.