Resource Folders for the MasterTree (#307)
* Fixed Dependecies

I moved some scenes in the last minute. Stupid me. I hope it is fixed now.

* Revert "Fixed Dependecies"

This reverts commit 9e7bcd8dbb29b69aa6373ad5553ebd686f437de6.

* Auto stash before merge of "main" and "upstream/main"

* Update Changelog

To contain the changes I made during the last week.

* Revert "Auto stash before merge of "main" and "upstream/main""

This reverts commit 8e1699d08a204b65e3f292067b2c4e870aaee1c7.

* Add the new styles to the remaining pickers

TimelinePicker, ThemePicker, OperationPicker, GlossaryPicker. It also removes the [] around the items, except for the operation picker. I wasn't sure if it wouldn't make readability harder having three similar buttons next to each other, so I kept the [] there. 

And it adds a "Bus:" label to the AudioPicker, because I thought it would make it cleaner.

* Added Explanation on some Dialogic classes

Also added sections so they are easier to search through.

I was basically trying to understand everything and wrote down what I found out for future contributors.

* Work on the Resource Folders

Right now folders and subfolders work for the timelines (because I wanted to get it working before I expand it). 

There is a new file called "folder_structure.json" in the dialogic folder. It is parsed to create the tree. 

Most of the functions to work with the folder structure are in the DialogicUtil.

I have moved the Popups into the MasterTree. They are now created in the script.
There is a new "RemoveFolderConfrimation" popup.

Also I changed the interaction between the MasterTree and the TimelineEditor for creating a new timeline. All the Tree stuff is now done in the tree ('new_timeline()') and only the create_timeline function of the TimelineEditor is called.

Next I want to try working on dragging and dropping resources. We'll see how that goes. After that there needs to be a lot of cleanup, because many things are not very nice right now. Wish me luck!

* MasterTree script cleanup

Just moved some code, added explanations, sections and stuff.

* Drag and Drop, better folder_structure update

You can now drag and drop the items in the MasterTree.

Also there is now a better folder_structure updater.

* Implement and polish ResourceFolders

ResourceFolders are now working for all 4 resource types. 

Like for the timelines, the Popupmenus have been expanded and the creation of new resources was shifted into the MasterTree.gd. 

Resources are now really deleted when one of their parent folders is deleted.

Folders now have metadata. Right now it contains a color key and a folded key. The second one is used to save the folded/unfolded state of each folder. The color key is not used right now. Metadata can easily be read and written with two easy functions in the DialogicUtil.

When moving, adding and renaming folders, it now makes sure that no folder folder whith that name already exists (as that produced errors).

The folder structure is now updated once when the editor is loaded to make sure there are no files missing.

I'm not sure what the problem was, but when editing the names in the CharacterEditor or the DefinitionEditor, the LineEdit lost focus each time you entered a character. This was fixed by just grabbing the focus back ;).

Btw, something I didn't explain in the previous commits: The master tree has some new functions that are used to get the path to an item in the tree. That path can be used to access for example the folder in the folder structure.

* Small Bug fixes

When dragging one of the root folders, the folder structure would become corrupted. Not anymore.

When editing a folder name but not changing it, it would corrupt the folder. Not anymore.

Turns out moving folders didn't work anymore, since it removes the old folder and then adds the new one. But when removing the old one it also REALLY deleted all the contained files. So each time you moved a folder it would delete all it's files. 

Took me long to find out and short to fix.

* Using the folder icon as in the native FileSystem tree

* Small bug fix for renaming folders

Same as the one for moving folders, but still I missed it. Shame on me ;(.

* Small bug fixes

You cannot drag folder onto their subfolders anymore.

New theme files are now added to the selected folder, instead of always being moved into the root "Themes" folder by a leftover line in the ThemeEditor.

* Fixing the fix

I should probably stop making so many commits.

My previous fix made it impossible to drag files into subfolders of their parent folder.... Stupid.

* Improving the search/filtering of the MasterTree

Now only the needed folders are shown. Also folding/unfolding folders is not saved when a search term is entered.

* Fix for the Definition/Glossary icon

When changing the type of the definition it would set old icons.

* adding the proper color to folders

Co-authored-by: Emilio Coppola <coppolaemilio@gmail.com>