Folders for the ResourcePickers (#312)
* Fixed Dependecies

I moved some scenes in the last minute. Stupid me. I hope it is fixed now.

* Revert "Fixed Dependecies"

This reverts commit 9e7bcd8dbb29b69aa6373ad5553ebd686f437de6.

* Auto stash before merge of "main" and "upstream/main"

* Update Changelog

To contain the changes I made during the last week.

* Revert "Auto stash before merge of "main" and "upstream/main""

This reverts commit 8e1699d08a204b65e3f292067b2c4e870aaee1c7.

* Add the new styles to the remaining pickers

TimelinePicker, ThemePicker, OperationPicker, GlossaryPicker. It also removes the [] around the items, except for the operation picker. I wasn't sure if it wouldn't make readability harder having three similar buttons next to each other, so I kept the [] there. 

And it adds a "Bus:" label to the AudioPicker, because I thought it would make it cleaner.

* Folders for the EventPart_TimelinePicker

I tried implementing folders into the timeline picker

* Folders for the DefinitionPicker

The Definition Icon can not easily be loaded right now, as I don't have access to the editor_reference as easy as with the TimelinePicker.

* Folders for the CharacterPicker

I also fixed the annoying message "This signal was already connected..."

* Folders for the ThemePicker and the GlossaryPicker

Also I added the DefinitionIcon and the CharacterIcon to be loaded directly. It works for now, but if I get the editor reference to work one day, I would like to have it get the icon from the MasterTree.