Add support for disabeling the event body when not needed
You can now call the "request_set_body_ebabled" signal in every event_part (that is the direct body or header of an event). By passing true or false as argument you can either enable and expand the body or disable and hide the body.

This commit also already implements this for the two events that needed it:
- SetGlossary
- SetBackground
# in both cases this is handled by the body. It could be handled by the header too, but who cares.