BackgroundMusic fixes
The Background music can now fade more then once (previously the old track kept playing and so no new music was faded in).

Fading now is done on an exponantial curve to be more linear. That might be confusing but somehow the decibel scale is logarithmic. This means more fluid transitions.

I still need to test things better though. Also a tranisition will still fail if the previous one is still in progress.