Add ability to disable event body (Empty Body Unfold Fix) (#317)
* Fixed Dependecies

I moved some scenes in the last minute. Stupid me. I hope it is fixed now.

* Revert "Fixed Dependecies"

This reverts commit 9e7bcd8dbb29b69aa6373ad5553ebd686f437de6.

* Auto stash before merge of "main" and "upstream/main"

* Update Changelog

To contain the changes I made during the last week.

* Revert "Auto stash before merge of "main" and "upstream/main""

This reverts commit 8e1699d08a204b65e3f292067b2c4e870aaee1c7.

* Add the new styles to the remaining pickers

TimelinePicker, ThemePicker, OperationPicker, GlossaryPicker. It also removes the [] around the items, except for the operation picker. I wasn't sure if it wouldn't make readability harder having three similar buttons next to each other, so I kept the [] there. 

And it adds a "Bus:" label to the AudioPicker, because I thought it would make it cleaner.

* Add support for disabeling the event body when not needed

You can now call the "request_set_body_ebabled" signal in every event_part (that is the direct body or header of an event). By passing true or false as argument you can either enable and expand the body or disable and hide the body.

This commit also already implements this for the two events that needed it:
- SetGlossary
- SetBackground
# in both cases this is handled by the body. It could be handled by the header too, but who cares.

* Make sure events with disabled bodies don't expand

It was still possible to expand the events with disabled bodies/expand_controls when using the fold/unfold in the toolbar.

* Scene Preview for the EventPart_BackgroundPreview 

If a preview for the selected background scene can be created, it will now be shown in the background.