Definitions rework: Values and GlossaryEntries (#339)
* Fixed Dependecies

I moved some scenes in the last minute. Stupid me. I hope it is fixed now.

* Revert "Fixed Dependecies"

This reverts commit 9e7bcd8dbb29b69aa6373ad5553ebd686f437de6.

* Auto stash before merge of "main" and "upstream/main"

* Update Changelog

To contain the changes I made during the last week.

* Revert "Auto stash before merge of "main" and "upstream/main""

This reverts commit 8e1699d08a204b65e3f292067b2c4e870aaee1c7.

* Add the new styles to the remaining pickers

TimelinePicker, ThemePicker, OperationPicker, GlossaryPicker. It also removes the [] around the items, except for the operation picker. I wasn't sure if it wouldn't make readability harder having three similar buttons next to each other, so I kept the [] there. 

And it adds a "Bus:" label to the AudioPicker, because I thought it would make it cleaner.

* Splitting Values and Glossary Entries

There are two new editors replacing the old "DefinitionsEditor": The ValueEditor and the GlossaryEntryEditor. 
You cannot convert them from one to the other anymore. You can create each type using the toolbar or the context menu. Everything else stayed the same except for some renaming in the scripts.

* Updating tooltip

Co-authored-by: Emilio Coppola <coppolaemilio@gmail.com>