Glossary theme improvements (#341)
* Fixed Dependecies

I moved some scenes in the last minute. Stupid me. I hope it is fixed now.

* Revert "Fixed Dependecies"

This reverts commit 9e7bcd8dbb29b69aa6373ad5553ebd686f437de6.

* Auto stash before merge of "main" and "upstream/main"

* Update Changelog

To contain the changes I made during the last week.

* Revert "Auto stash before merge of "main" and "upstream/main""

This reverts commit 8e1699d08a204b65e3f292067b2c4e870aaee1c7.

* Add the new styles to the remaining pickers

TimelinePicker, ThemePicker, OperationPicker, GlossaryPicker. It also removes the [] around the items, except for the operation picker. I wasn't sure if it wouldn't make readability harder having three similar buttons next to each other, so I kept the [] there. 

And it adds a "Bus:" label to the AudioPicker, because I thought it would make it cleaner.

* Adding Font selection and Color selection for the glossary

And it works!

* Fixing Glossary Preview in the ThemeEditor

Previously it didn't move with the camera and the background was hidden. Now it isn't.

* Add Glossary BackgroundPanel option

You can now select a Stylebox to be used for the background of the glossary info. There is a new example asset stylebox that is used by default.

* Removing the old BackgroundColor of the glossary box

* using default color values instead of modulates and remove previous self_modulate from scene

Co-authored-by: Emilio Coppola <coppolaemilio@gmail.com>