Update the Docs design to better match the godot docs
I replaced the font with Noto Sans (the default godot editor font).

I implemented most of the colors (just used the accent_color and lightened it a bit for some elements).

I implemented some hacky solution for the spacing on the left and the right. Maybe I'll find a better way one day.