Builtin Documentation (#278)
* Added DocumentationEditor

It's not like it's already working. Just like to go in small steps

* Implemented a simple Documentation System

It's still in the work though...
There is now a DocumentationEditor (more like a viewer actually).
And a Documentation folder that contains the docs.

* Renamed DocumentationEditor to Documentation Viewer and added Fonts

* Tried creating a documentation feature that works on it's own.

* Moved the images

* Updating the DocumentationContent for testing

* New way of integrating the docs

There now exist a couple of nodes to use in your plugin:
- DocsPageViewer (A RichTextLabel that can parse Markdown and handle links, scroll to sections, etc.)
- DocumentationTree (A tree with all the items of the Docs Content folder)
- DocumentationViewer (A working combination of the DocsPageViewer and the DocumentationTree)
You can now just use the DocumentationViewer as a kind of Documentation tab or page in your plugin. Or you can integrate the DocsPageViewer into an exisiting editor. It is also very easy to add the tree items to an existing tree.

This commit also worked on to the integration into dialogic, but I try to keep it as minimal as possible.

* Remove Spaces in the Images Names

* Tried editing the docs on github.

* Update Start.md

* Testing github links

* Update the content

* Updated internal Documentation Scripts and Scenes

* Updated the Docs integration

Searching is now possible.

* Trying github imagelinks

* Redoing the test

* Updating some images

* Use title (first line) instead of name in the tree

* Rename Tutorials and References

* Update Link

* Added Editing features

* Edited Content

* Update Exporting.md

* Add support for relative paths for images

* Testing relative paths

* quick rename of current_page to current_path and adding the ability to have relative paths as links

* Updating Test Content

* Add an Up Button and a Content Menu

The design is not perfect but it works. I think most pages should be relatively short.

* Updated the Content; added FAQ

* Fixed some of the markdown parser

One of the fixes is disabeling the *italic* part. It didn't work. Have to fix this one time.

* Rename Scripts Folder to Nodes

Just makes more sense.

* Updated the content. Added needed Images to the Step by Step guide

* Updated Regex for list.

Previously you couldn't use a + inside a text. That was annoying.

* Update the styling of the content menu and the page viewer

* Updated Content, Added Events Section to the reference

* Updated the Mono Font

* Removed print signal (#258)

* Small Changes

Removed the unnecessary scripts

* Testing the docs

* making some edits on the icons

* Remove some loops from the Markdownparser

Many loops were unnecessary.

* And fixed some things that were broken from the last commit...

* Removed the DocsHelper singlton

I added a DocsTreeHelper Node, that has to added to the Master Tree to access it. Also now the path to the documentation is more hardcoded in. I don't think this should be a problem though.

* Make the editor work again

No clue how, but the instance link was messed up with the GlossaryItemEditor. You can view pages again now.

* Fix the contents menu

* Added Documentation Context Menus and Better Content

* Fix an indentation error

* Update the Docs design to better match the godot docs

I replaced the font with Noto Sans (the default godot editor font).

I implemented most of the colors (just used the accent_color and lightened it a bit for some elements).

I implemented some hacky solution for the spacing on the left and the right. Maybe I'll find a better way one day.

* General Improvements/Fixes

Scrolling and jumping to sections works again, but I had to redo the whole spacing thing...

I added an option to enable the editing tools to the context menu.

Changed the Tree icons.

Changed some content.

* Added <> to have a visited history.

You have now two arrows in the upper left corner. You can use them to go back to previously visited pages.

Also I changed the welcome image to not only include documentation.

* Added a system to link documentation pages to event blocks

You can now specify a documentation page path in the event blocks settings (it's an exported variable). 

If one is specified it will show the help icon and clicking it will bring you to the page.

* Add the character join reference.

* Make Font Size a bit responsive

The basic fonts are now relative to get_constant("scale", "Editor").

Not sure how to work with the headings yet.

Co-authored-by: Emilio Coppola <coppolaemilio@gmail.com>