UndoRedo support for the TimelineEditor (#428)
* Fixed Dependecies

I moved some scenes in the last minute. Stupid me. I hope it is fixed now.

* Revert "Fixed Dependecies"

This reverts commit 9e7bcd8dbb29b69aa6373ad5553ebd686f437de6.

* Auto stash before merge of "main" and "upstream/main"

* Update Changelog

To contain the changes I made during the last week.

* Revert "Auto stash before merge of "main" and "upstream/main""

This reverts commit 8e1699d08a204b65e3f292067b2c4e870aaee1c7.

* Add the new styles to the remaining pickers

TimelinePicker, ThemePicker, OperationPicker, GlossaryPicker. It also removes the [] around the items, except for the operation picker. I wasn't sure if it wouldn't make readability harder having three similar buttons next to each other, so I kept the [] there. 

And it adds a "Bus:" label to the AudioPicker, because I thought it would make it cleaner.

* Cleanup and Simplification of the MasterTree.gd

* Auto stash before merge of "main" and "upstream/main"

* Remove more Engine.get_singleton()'s

:(

* Make the DocsTreeHelper ignore the mouse

It was blocking a small part of the MasterTree. This should fix #414

* Add UndoRedo support for some actions in the TimelineEditor

Should work for the following:

- deleting events
- cutting events
- pasting events
- moving events by dragging them

* Updated the changelog a bit

* Support CRLT+Y for Redo

* Duplicate events and adding through buttons

are now udnoable.

I'm not sure but there might be something wrong with the undo and redo of duplication...

Next thing is the adding through dropping ;(

* Make Undo Redo of dragged in events work

* Add Undo support for CRTL+T

Also the undo history only will be cleared when really changing timelines now.

* Fixing indentation issues generated from resolving conflicts

Co-authored-by: Emilio Coppola <coppolaemilio@gmail.com>