Documentation Fixes, Addition of Audio Events page
I fixed the broken Doc links in the events and also some the link-regex, that got confused by other [] in the text. 

I added a Audio Events page for both the Audio Event and the Background Music Event.

I added a note in the Text Editor docs about the BBcode options.