Documentation Additions/Fixes (#479)
* Documentation Fixes, Addition of Audio Events page

I fixed the broken Doc links in the events and also some the link-regex, that got confused by other [] in the text. 

I added an Audio Events page for both the Audio Event and the Background Music Event.

I added a note in the Text Editor docs about the BBcode options.

* Removed # check inside headings

Should fix #468

* #-Fix addition

Found out why it was in there before XD. But I found a better way to do it.