TimelineEditor: Scroll to newly created events (#486)
* Fixed Dependecies

I moved some scenes in the last minute. Stupid me. I hope it is fixed now.

* Revert "Fixed Dependecies"

This reverts commit 9e7bcd8dbb29b69aa6373ad5553ebd686f437de6.

* Auto stash before merge of "main" and "upstream/main"

* Update Changelog

To contain the changes I made during the last week.

* Revert "Auto stash before merge of "main" and "upstream/main""

This reverts commit 8e1699d08a204b65e3f292067b2c4e870aaee1c7.

* Add the new styles to the remaining pickers

TimelinePicker, ThemePicker, OperationPicker, GlossaryPicker. It also removes the [] around the items, except for the operation picker. I wasn't sure if it wouldn't make readability harder having three similar buttons next to each other, so I kept the [] there. 

And it adds a "Bus:" label to the AudioPicker, because I thought it would make it cleaner.

* Cleanup and Simplification of the MasterTree.gd

* Auto stash before merge of "main" and "upstream/main"

* Remove more Engine.get_singleton()'s

:(

* Make the DocsTreeHelper ignore the mouse

It was blocking a small part of the MasterTree. This should fix #414

* VoiceSupport Feature Fix: Make sure no default bus is needed

There was a little annyoing warning popping up a lot.

* Small Documentation fix

* Add event documentation

Adds a basic documentation of
- Leave Event
- Question Event
- Choice Event
- End Event
- Condition Event
- Set Value Event

* Add the new documentations to the events

You now have these little buttons on the right of the events that instantly lead to the documentation.

* Scrolling to newly created events

* only scrolling to pieces created by clicking on the button

Co-authored-by: Emilio Coppola <coppolaemilio@gmail.com>