In in a nw=v feature
With this, the text event will skip to the next, the moment the voiceacting is complete.
It is a hack, and needs to be better refined to fit in base code, but it works.

Tests work as expected.
When tested on a line without voice, it defaults to a 0 secund skip