Better handling for custom events. Prevents orphan nodes to be created when handling it.
* DialogicCustomEvents.gd:
This script is attached to a Node (called "CustomEvents") that is a child of the DialogNode.
It updates the custom events handlers by searching the custom events folder (just like before) and creating a node for every event.
Keeps the data of the handlers node in a dictionary (handlers) with keys being the event_ids and values being a the handler node reference