New TextEvent commands [signal], [pause], [play], [nw=v]
* In in a nw=v feature

With this, the text event will skip to the next, the moment the voiceacting is complete.
It is a hack, and needs to be better refined to fit in base code, but it works.

Tests work as expected.
When tested on a line without voice, it defaults to a 0 secund skip

* Test proof of concept for tracking location of a text command

Test succesful.
The hello thing console output appeared where expected, is possebly one character off. This may be related to rendering due to use of a breakpoint to test.

* Implement fundamental text in commands

Currently untested, you're welcome future me.

Added a univerial regex that should fill all current and planned uses of commands in the text bubble.
Rewrote most of update_text to use the new regex.
Lots of comments for documentation
Add func handle_event to handle events as they trigger

* Complete proof of concept

Basic tests completed, all features working as expected.
One hacky bypass included, must be fixed before merging.

* Adding a Pause command

* update the regex doc comments

* Attempt to re-implement play command

Problem: the new audiostreamplayer is not making any sound. I am unable to test if the thing is working.

* Update Changelog and Documentation

Co-authored-by: KvaGram <0kvakk@gmailcom>
Co-authored-by: Jowan-Spooner <raban-loeffler@posteo.de>