Fixed some scrolling issues and inconsistencies
There were some troubles getting the RichTextLabel to co-operate with
its own constantly-changing size. This should hopefully keep things
consistent even with long messages. I also mistakenly applied the margin
incorrectly to the portrait, causing it to add extra padding on the
wrong side when it was displayed on the text box. This is now fixed. A
good future goal would be to dynamically move the "next message" prompt
when it overlaps with the portrait frame.

remove debug print statement